refactor: package를 이용한 layer 분리
