+++ src/main/java/step3/application/SimulationCondition.java
@@ -1,0 +1,20 @@
+package step3.application;
+
+public class SimulationCondition {
+    private final int numberOfCar;
+    private final int numberOfAttempts;
+
+    public SimulationCondition(final int numberOfCar, final int numberOfAttempts) {
+        //TODO:: validation 추가
+        this.numberOfCar = numberOfCar;
+        this.numberOfAttempts = numberOfAttempts;
+    }
+
+    public int getNumberOfCar() {
+        return numberOfCar;
+    }
+
+    public int getNumberOfAttempts() {
+        return numberOfAttempts;
+    }
+}
