+++ src/main/java/step3/domain/Snapshot.java
@@ -1,0 +1,19 @@
+package step3.domain;
+
+public class Snapshot {
+    private final int number;
+    private final RacingMap racingMap;
+
+    public Snapshot(final int number, final RacingMap racingMap) {
+        this.number = number;
+        this.racingMap = racingMap.clone();
+    }
+
+    public int getNumber() {
+        return number;
+    }
+
+    public RacingMap getRacingMap() {
+        return racingMap;
+    }
+}
