+++ src/main/java/step3/interfaces/output/Printer.java
@@ -1,0 +1,11 @@
+package step3.interfaces.output;
+
+public class Printer {
+    public static void println(final Object x) {
+        System.out.println(x);
+    }
+
+    public static void print(final Object x) {
+        System.out.print(x);
+    }
+}
